chicken-diary-app/app/requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.7
Flask-SQLAlchemy==3.0.5
flask-migrate==4.0.7
python-dotenv
